pyproject.toml
[project]
name = "carekobot"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "nayetdet", email = "jpmoreiracampos21@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "discord-py[voice]>=2.6.3",
    "python-dotenv>=1.1.1",
]

[tool.uv.workspace]
members = []
